Some test refactoring
